+++ src/aql.ts
@@ -1,20 +1,52 @@
 import { ArangoCollection, isArangoCollection } from "./collection";
+import { Dict } from "./util/types";
 
+/**
+ * Generic AQL query object consisting of an AQL query string and its bind
+ * parameters.
+ */
 export interface AqlQuery {
+  /**
+   * An AQL query string.
+   */
   query: string;
-  bindVars: { [key: string]: any };
+  /**
+   * An object mapping AQL bind parameter names to their respective values.
+   *
+   * Names of parameters representing collections are prefixed with an
+   * at-symbol.
+   */
+  bindVars: Dict<any>;
 }
 
-interface GeneratedAqlQuery extends AqlQuery {
+/**
+ * Derived type representing AQL query objects generated by the AQL helper
+ * functions and the AQL template string handler. These objects can be fed
+ * back into these helper functions to be inlined or merged in complex queries.
+ *
+ * @internal
+ */
+export interface GeneratedAqlQuery extends AqlQuery {
+  /** @hidden @internal */
   _source: () => { strings: string[]; args: AqlValue[] };
 }
 
+/**
+ * An object representing a trusted AQL literal that will be inlined directly
+ * when used in an AQL template or passed to AQL helper functions.
+ *
+ * Arbitrary values can be converted to trusted AQL literals by passing them
+ * to the {@link aql.literal} helper function.
+ */
 export interface AqlLiteral {
-  /** @hidden */
+  /** @hidden @internal */
   toAQL: () => string;
 }
 
-/** @hidden */
+/**
+ * A value that can be used in an AQL template string or passed to an AQL
+ * helper function.
+ */
 export type AqlValue =
   | ArangoCollection
   | GeneratedAqlQuery
@@ -31,7 +63,8 @@ export function isAqlQuery(query: any): query is AqlQuery {
   return Boolean(query && typeof query.query === "string" && query.bindVars);
 }
 
-function isGeneratedAqlQuery(query: any): query is GeneratedAqlQuery {
+/** @hidden @internal */
+export function isGeneratedAqlQuery(query: any): query is GeneratedAqlQuery {
   return isAqlQuery(query) && typeof (query as any)._source === "function";
 }
