+++ src/lib/multipart.web.ts
@@ -1,5 +1,5 @@
+import { Errback } from "../util/types";
 import { Fields, MultipartRequest } from "./multipart";
-import { Errback } from "./types";
 
 declare const FormData: any;
 declare const Blob: any;
