+++ src/test/14-document-collections.ts
@@ -1,6 +1,7 @@
 import { expect } from "chai";
-import { DocumentCollection, DocumentMetadata } from "../collection";
+import { DocumentCollection } from "../collection";
 import { Database } from "../database";
+import { DocumentMetadata } from "../documents";
 
 const ARANGO_URL = process.env.TEST_ARANGODB_URL || "http://localhost:8529";
 const ARANGO_VERSION = Number(
